requirements.txt
lifxlan==1.2.5
python-daemon